services/inference-engine/inference_engine.py
import os
import cv2
import numpy as np
import logging
from typing import Dict, Tuple, List
from threading import Thread
from queue import Queue
from ultralytics import YOLO
from insightface.app import FaceAnalysis
from prometheus_client import Counter, Gauge, start_http_server
import paho.mqtt.client as mqtt
import redis

class ModelManager:
    def __init__(self, model_path: str):
        self.model_path = model_path
        self.models = self._load_models()
        self.metrics = {
            'inference_time': Gauge('inference_time', 'Model inference time in ms'),
            'detection_count': Counter('detection_count', 'Number of faces detected')
        }
        
    def _load_models(self) -> Dict:
        # Load YOLOv8-Face model with NPU acceleration
        yolo_model = YOLO(f"{self.model_path}/yolov8n-face.pt")
        yolo_model.to('cpu')  # Will use NPU via TFLite delegate
        
        # Load InsightFace model for detection and recognition
        app = FaceAnalysis(providers=['CPUExecutionProvider'])
        app.prepare(ctx_id=0, det_size=(640, 640))
        
        return {
            'yolo': yolo_model,
            'insightface': app
        }

class FaceProcessor:
    def __init__(self, model_manager: ModelManager):
        self.model_manager = model_manager
        self.face_db = redis.Redis(host='redis', port=6379, db=0)
        
    def detect_faces(self, frame: np.ndarray) -> List[Dict]:
        # Hybrid detection using YOLOv8 and InsightFace
        yolo_results = self.model_manager.models['yolo'](frame)
        insight_results = self.model_manager.models['insightface'].get(frame)
        
        # Merge and filter results
        faces = self._merge_detections(yolo_results, insight_results)
        return faces
    
    def recognize_face(self, face_img: np.ndarray) -> str:
        # Extract features using InsightFace
        face_features = self.model_manager.models['insightface'].get(face_img)
        if not face_features:
            return None
        features = face_features[0].embedding
        
        # Search in Redis for matching face
        matches = self._search_face_db(features)
        return matches[0] if matches else None
    
    def check_liveness(self, face_img: np.ndarray, frame_sequence: List[np.ndarray]) -> bool:
        # Texture analysis
        lbp_score = self._analyze_texture(face_img)
        
        # Color space analysis
        hsv_score = self._analyze_color_space(face_img)
        
        # Motion analysis
        if len(frame_sequence) > 1:
            flow_score = self._analyze_optical_flow(frame_sequence)
            blink_score = self._detect_blink(frame_sequence)
        else:
            flow_score = 1.0
            blink_score = 1.0
            
        # Combine scores
        final_score = (lbp_score + hsv_score + flow_score + blink_score) / 4
        return final_score > 0.8
    
    def _merge_detections(self, yolo_results, retina_results) -> List[Dict]:
        # Implement NMS and confidence-based merging
        merged = []
        # ... (implementation details)
        return merged
    
    def _analyze_texture(self, face_img: np.ndarray) -> float:
        # Implement Local Binary Pattern analysis
        # ... (implementation details)
        return 0.95
    
    def _analyze_color_space(self, face_img: np.ndarray) -> float:
        # Analyze HSV color space for skin tone consistency
        # ... (implementation details)
        return 0.90
    
    def _analyze_optical_flow(self, frame_sequence: List[np.ndarray]) -> float:
        # Calculate dense optical flow between frames
        # ... (implementation details)
        return 0.85
    
    def _detect_blink(self, frame_sequence: List[np.ndarray]) -> float:
        # Implement blink detection using facial landmarks
        # ... (implementation details)
        return 0.88

class InferenceEngine:
    def __init__(self):
        self.model_manager = ModelManager('/app/models')
        self.face_processor = FaceProcessor(self.model_manager)
        self.frame_queue = Queue(maxsize=100)
        self.mqtt_client = self._setup_mqtt()
        start_http_server(8000)  # Metrics endpoint
        
    def _setup_mqtt(self) -> mqtt.Client:
        client = mqtt.Client()
        client.on_message = self._on_frame
        client.connect('mqtt-broker', 1883)
        client.subscribe('frames/#')
        return client
    
    def _on_frame(self, client, userdata, message) -> None:
        # Decode frame from MQTT message
        nparr = np.frombuffer(message.payload, np.uint8)
        frame = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        self.frame_queue.put((message.topic, frame))
    
    def start(self) -> None:
        # Start processing thread
        self.process_thread = Thread(target=self._process_frames)
        self.process_thread.daemon = True
        self.process_thread.start()
        
        # Start MQTT loop
        self.mqtt_client.loop_forever()
    
    def _process_frames(self) -> None:
        while True:
            topic, frame = self.frame_queue.get()
            
            # Process frame
            faces = self.face_processor.detect_faces(frame)
            
            for face in faces:
                # Extract face image
                face_img = frame[face['bbox'][1]:face['bbox'][3],
                                face['bbox'][0]:face['bbox'][2]]
                
                # Check liveness
                if self.face_processor.check_liveness(face_img, [frame]):
                    # Recognize face
                    person_id = self.face_processor.recognize_face(face_img)
                    
                    if person_id:
                        # Publish recognition result
                        result = {
                            'person_id': person_id,
                            'confidence': face['confidence'],
                            'timestamp': time.time()
                        }
                        self.mqtt_client.publish(
                            f'recognition/{topic.split("/")[1]}',
                            json.dumps(result)
                        )

if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)
    engine = InferenceEngine()
    engine.start()