services/inference-engine/requirements.txt
# Deep Learning Frameworks
tflite-runtime
onnxruntime

# Computer Vision
opencv-python-headless
insightface

# AI Models
ultralytics
insightface>=0.7

# Anti-spoofing
scikit-image
scikit-learn

# Message Broker
paho-mqtt
redis

# Monitoring and Logging
prometheus-client
structlog

# Utilities
numpy
pyyaml
cryptography

# Graphics Libraries
pycairo
pygobject